+++ src/main/java/io/github/poshjosh/ratelimiter/web/core/RequestInfos.java
@@ -1,0 +1,9 @@
+package io.github.poshjosh.ratelimiter.web.core;
+
+import javax.servlet.http.HttpServletRequest;
+
+public interface RequestInfos {
+    static RequestInfo of(HttpServletRequest request) {
+        return new DefaultRequestInfo(request);
+    }
+}