+++ src/main/java/io/github/poshjosh/ratelimiter/web/core/WebExpressionMatchers.java
@@ -3,10 +3,8 @@ package io.github.poshjosh.ratelimiter.web.core;
 import io.github.poshjosh.ratelimiter.expression.*;
 import javax.servlet.http.HttpServletRequest;
 
-public interface WebExpressionMatcher
-        extends ExpressionMatcher<HttpServletRequest>, WebExpressionKey{
-
-    static WebExpressionMatcher ofHttpServletRequest() {
+public interface WebExpressionMatchers {
+    static ExpressionMatcher<HttpServletRequest> ofHttpServletRequest() {
         return new DefaultWebExpressionMatcher();
     }
 }