+++ src/main/java/io/github/poshjosh/ratelimiter/web/core/util/DefaultResourceInfo.java
@@ -2,14 +2,14 @@ package io.github.poshjosh.ratelimiter.web.core.util;
 
 import java.util.*;
 
-final class DefaultResourceInfo implements ResourceInfoProvider.ResourceInfo {
+final class DefaultResourceInfo implements ResourceInfo {
     private final Collection<String> httpMethods;
-    private final PathPatterns<String> pathPatterns;
+    private final ResourcePath<String> resourcePath;
     private final String id;
-    DefaultResourceInfo(PathPatterns<String> pathPatterns, String...httpMethods) {
+    DefaultResourceInfo(ResourcePath<String> resourcePath, String...httpMethods) {
         this.httpMethods = toCollection(httpMethods);
-        this.pathPatterns = Objects.requireNonNull(pathPatterns);
-        final List<String> paths = getPathPatterns().getPatterns();
+        this.resourcePath = Objects.requireNonNull(resourcePath);
+        final List<String> paths = getResourcePath().getPatterns();
         this.id = paths.isEmpty() ? this.httpMethods.toString() : this.httpMethods.toString() + paths;
     }
     private Collection<String> toCollection(String...arr) {
@@ -18,17 +18,17 @@ final class DefaultResourceInfo implements ResourceInfoProvider.ResourceInfo {
     }
     @Override public String getId() { return id; }
     @Override public Collection<String> getHttpMethods() { return httpMethods; }
-    @Override public PathPatterns<String> getPathPatterns() { return pathPatterns; }
+    @Override public ResourcePath<String> getResourcePath() { return resourcePath; }
     @Override public boolean equals(Object o) {
         if (this == o)
             return true;
         if (o == null || getClass() != o.getClass())
             return false;
         DefaultResourceInfo that = (DefaultResourceInfo) o;
-        return httpMethods.equals(that.httpMethods) && pathPatterns.equals(that.pathPatterns);
+        return httpMethods.equals(that.httpMethods) && resourcePath.equals(that.resourcePath);
     }
     @Override public int hashCode() {
-        return Objects.hash(httpMethods, pathPatterns);
+        return Objects.hash(httpMethods, resourcePath);
     }
     @Override public String toString() { return "ResourceInfo{" + id + '}'; }
 }