+++ src/main/java/io/github/poshjosh/ratelimiter/web/core/util/ResourceInfo.java
@@ -1,0 +1,12 @@
+package io.github.poshjosh.ratelimiter.web.core.util;
+
+import java.util.Collection;
+
+public interface ResourceInfo {
+
+    String getId();
+
+    ResourcePath<String> getResourcePath();
+
+    Collection<String> getHttpMethods();
+}