+++ src/test/java/io/github/poshjosh/ratelimiter/web/core/WebRateLimiterRegistryTest.java
@@ -37,6 +37,6 @@ class WebRateLimiterRegistryTest {
                 .resourceInfoProvider(ResourceInfoProvider.NONE)
                 .classes(this.getClass())
                 .build();
-        return WebRateLimiterRegistry.of(config);
+        return WebRateLimiterRegistries.of(config);
     }
 }